Moves stock placeholder model to placeholder package
